MultiAgentAssist/requirements.txt
# --- Interface Graphique (UI) ---
kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.8.0

# --- Intelligence Artificielle & Réseau ---
openai>=1.0.0
certifi

# --- Scraping & Data ---
youtube-search-python
numpy

# --- Utilitaires ---
requests